# docs/openapi/swagger.yaml
consumes:
- application/json
definitions:
...
  User:
    description: User is the generic user type
    properties:
      created_at:
        description: Date when the User was added
        example: "2020-09-17T03:07:13.418204+05:30"
        format: date-time
        type: string
        x-go-name: CreatedAt
      email:
        description: User email
        example: jdoe@gmail.com
        type: string
        x-go-name: Email
      id:
        description: User ID
        example: 24
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: User name
        example: John Doe
        type: string
        x-go-name: Name
      role:
        description: User role (admin | developer)
        example: developer
        type: string
        x-go-name: Role
      updated_at:
        description: Date when the User was updated
        example: "2020-09-17T03:07:13.418204+05:30"
        format: date-time
        type: string
        x-go-name: UpdatedAt
      url:
        description: User website
        example: https://jdoe.co
        type: string
        x-go-name: URL
      username:
        description: User username
        example: jdoe
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/PrunedNeuron/Fluoride/pkg/model
host: localhost:3000
info:
  contact:
    email: am@ayushm.dev
    name: Ayush Mishra
    url: http://ayushm.dev
  description: This is the API documentation page for the backend API used by the fluoride icon pack management service. Feel free to try hitting the endpoints using the interactive interface.
  license:
    name: GPL 3
    url: https://opensource.org/licenses/gpl-3.0
  title: Fluoride, a robust icon pack management service
  version: 0.0.1
paths:
  /developers:
    get:
      operationId: GetDevs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              developers:
                items:
                  $ref: '#/definitions/User'
                type: array
              status:
                description: status message
                example: success
                type: string
            type: object
      summary: Gets the list of all the developers present in the database.
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Saves a new developer to the database, rejecting the request on conflict
      operationId: SaveDev
      parameters:
      - description: information about the new developer
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              description: email address of the developer
              example: am@ayushm.dev
              type: string
            name:
              description: name of the developer
              example: Ayush Mishra
              type: string
            role:
              description: role of the user (developer | admin)
              example: developer
              type: string
            url:
              description: developer website URL
              example: https://ayushm.dev
              type: string
            username:
              description: username of the developer
              example: ayush
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                description: informational message
                example: successfully added developer
                type: string
              status:
                description: status message
                example: success
                type: string
            type: object
        "500":
          description: server error
          schema:
            properties:
              error:
                description: error message
                example: user may already exist
                type: string
              message:
                description: informational message
                example: invalid request
                type: string
              status:
                description: status message
                example: failure
                type: string
            type: object
      security:
      - api_key: []
      summary: Add a developer
      tags:
      - Users
  /developers/{developer}:
    get:
      description: Fetches the developer with the provided username
      operationId: GetDevByUsername
      parameters:
      - description: developer username
        in: path
        name: developer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              developers:
                items:
                  $ref: '#/definitions/User'
                type: array
              status:
                description: status message
                example: success
                type: string
            type: object
        "500":
          description: bad request
          schema:
            properties:
              error:
                description: error message
                example: invalid dev
                type: string
              message:
                description: informational message
                example: server error
                type: string
              status:
                description: status message
                example: failure
                type: string
            type: object
      summary: Get a developer by username
      tags:
      - Users
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  api_key:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"